+++ src/txstack.h
@@ -51,8 +51,8 @@ class Stack {
     std::vector<StackSlot> Table;
 
 private:
-    size_t                 last_xid; // id of the last
-    size_t                 xid_boot;
+    long                   last_xid; // id of the last
+    long                   xid_boot;
     Istring                cur_lid;    // the id to be pushed on uids[]
     std::vector<AttList>   attributes; // the main table of attributes
     std::vector<Xml *>     enames;     // the main table of element names
